+++ segments/pass/pass.go
@@ -1,6 +1,6 @@
 // Serves as a template for new segments and forwards flows, otherwise does
 // nothing.
-package noop
+package pass
 
 import (
 	"sync"
@@ -8,18 +8,18 @@ import (
 	"github.com/bwNetFlow/flowpipeline/segments"
 )
 
-// The NoOp Segment is considered a template for any additional Segments, as it
+// The Pass Segment is considered a template for any additional Segments, as it
 // showcases the exact implementation.
-type NoOp struct {
+type Pass struct {
 	segments.BaseSegment // always embed this, no need to repeat I/O chan code
 	// add any additional fields here
 }
 
 // Every Segment must implement a New method, even if there isn't any config
 // it is interested in.
-func (segment NoOp) New(config map[string]string) segments.Segment {
+func (segment Pass) New(config map[string]string) segments.Segment {
 	// do config stuff here, add it to fields maybe
-	return &NoOp{}
+	return &Pass{}
 }
 
 // The main goroutine of any Segment. Any Run method must:
@@ -30,7 +30,7 @@ func (segment NoOp) New(config map[string]string) segments.Segment {
 // Usually, when using a range over In in combination with below defer, nothing
 // will go wrong. However, some segments have a legitimate use case for using
 // `for {}`, in which case care must be taken to keep draining In.
-func (segment *NoOp) Run(wg *sync.WaitGroup) {
+func (segment *Pass) Run(wg *sync.WaitGroup) {
 	defer func() {
 		// This defer clause is important and needs to be present in
 		// any Segment.Run method in some form, but with at least the
@@ -48,6 +48,6 @@ func (segment *NoOp) Run(wg *sync.WaitGroup) {
 // callable from config. An unregistered Segment will only be available using
 // the API.
 func init() {
-	segment := &NoOp{}
-	segments.RegisterSegment("noop", segment)
+	segment := &Pass{}
+	segments.RegisterSegment("paspass", segment)
 }
